bab master data selesai

+++ trunk/java-desktop-book/code/src/com/googlecode/projecttemplate/pos/dao/BaseDaoHibernate.java
@@ -23,7 +23,7 @@ public class BaseDaoHibernate<T> {
         return domain;
     }
     @SuppressWarnings("unchecked")
-    public T getById(Long id) {
+    public T getById(Integer id) {
         return (T) sessionFactory.getCurrentSession().get(domainClass, id);
     }
     public T delete(T domain) {
